Added a method for the bot to save the chatid of each individual player

+++ src/Player.java
@@ -3,9 +3,9 @@ public class Player {
 	public String role;
 	public boolean isKing;
 	public String name;
-	public int id;
+	public long id;
 	
-	public Player(int id, String name) {
+	public Player(long id, String name) {
 		this.id = id;
 		this.name = name;
 	}
